separate NewDocumentViewController
